start moving files
