+++ public/index.html
@@ -29,14 +29,7 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div class="page">
-
-        <div id="header">
-          <div id="App"></div>
-        </div>
-        <div id="side"> </div>
-        <div id="map"></div>
-    </div>
+    <div id="root"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
