+++ src/App.js
@@ -45,7 +45,7 @@ const App = () => (
     <BaseLayout />
   </BrowserRouter>
 )
-// render(<App />, document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
 
 
 export default App;
