+++ src/components/Side/side.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Input from './Input/input.js';
+
+export default function Side() {
+    const side = (
+      <Input/>
+    );
+}
