+++ src/components/header/header.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export default function Header() {
+    return(
+      <div class="header">
+        <h1>Tokyo coffee map - with wifi & plug.</h1>
+          <a href="https://github.com/theyij/tokyo-dev-map" rel="noopener
+   noreferrer" target="_blank">
+            <p>github</p>
+          </a>
+      </div>
+    );
+}
